+++ pqcrypto-kyber/src/ffi.rs
@@ -17,73 +17,169 @@ pub const PQCLEAN_KYBER512_CLEAN_CRYPTO_SECRETKEYBYTES: usize = 1632;
 pub const PQCLEAN_KYBER512_CLEAN_CRYPTO_PUBLICKEYBYTES: usize = 800;
 pub const PQCLEAN_KYBER512_CLEAN_CRYPTO_CIPHERTEXTBYTES: usize = 736;
 pub const PQCLEAN_KYBER512_CLEAN_CRYPTO_BYTES: usize = 32;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER512_AVX2_CRYPTO_SECRETKEYBYTES: usize = 1632;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER512_AVX2_CRYPTO_PUBLICKEYBYTES: usize = 800;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER512_AVX2_CRYPTO_CIPHERTEXTBYTES: usize = 736;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER512_AVX2_CRYPTO_BYTES: usize = 32;
 pub const PQCLEAN_KYBER768_CLEAN_CRYPTO_SECRETKEYBYTES: usize = 2400;
 pub const PQCLEAN_KYBER768_CLEAN_CRYPTO_PUBLICKEYBYTES: usize = 1184;
 pub const PQCLEAN_KYBER768_CLEAN_CRYPTO_CIPHERTEXTBYTES: usize = 1088;
 pub const PQCLEAN_KYBER768_CLEAN_CRYPTO_BYTES: usize = 32;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER768_AVX2_CRYPTO_SECRETKEYBYTES: usize = 2400;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER768_AVX2_CRYPTO_PUBLICKEYBYTES: usize = 1184;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER768_AVX2_CRYPTO_CIPHERTEXTBYTES: usize = 1088;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER768_AVX2_CRYPTO_BYTES: usize = 32;
 pub const PQCLEAN_KYBER1024_CLEAN_CRYPTO_SECRETKEYBYTES: usize = 3168;
 pub const PQCLEAN_KYBER1024_CLEAN_CRYPTO_PUBLICKEYBYTES: usize = 1568;
 pub const PQCLEAN_KYBER1024_CLEAN_CRYPTO_CIPHERTEXTBYTES: usize = 1568;
 pub const PQCLEAN_KYBER1024_CLEAN_CRYPTO_BYTES: usize = 32;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER1024_AVX2_CRYPTO_SECRETKEYBYTES: usize = 3168;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER1024_AVX2_CRYPTO_PUBLICKEYBYTES: usize = 1568;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER1024_AVX2_CRYPTO_CIPHERTEXTBYTES: usize = 1568;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER1024_AVX2_CRYPTO_BYTES: usize = 32;
 pub const PQCLEAN_KYBER51290S_CLEAN_CRYPTO_SECRETKEYBYTES: usize = 1632;
 pub const PQCLEAN_KYBER51290S_CLEAN_CRYPTO_PUBLICKEYBYTES: usize = 800;
 pub const PQCLEAN_KYBER51290S_CLEAN_CRYPTO_CIPHERTEXTBYTES: usize = 736;
 pub const PQCLEAN_KYBER51290S_CLEAN_CRYPTO_BYTES: usize = 32;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER51290S_AVX2_CRYPTO_SECRETKEYBYTES: usize = 1632;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER51290S_AVX2_CRYPTO_PUBLICKEYBYTES: usize = 800;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER51290S_AVX2_CRYPTO_CIPHERTEXTBYTES: usize = 736;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER51290S_AVX2_CRYPTO_BYTES: usize = 32;
 pub const PQCLEAN_KYBER76890S_CLEAN_CRYPTO_SECRETKEYBYTES: usize = 2400;
 pub const PQCLEAN_KYBER76890S_CLEAN_CRYPTO_PUBLICKEYBYTES: usize = 1184;
 pub const PQCLEAN_KYBER76890S_CLEAN_CRYPTO_CIPHERTEXTBYTES: usize = 1088;
 pub const PQCLEAN_KYBER76890S_CLEAN_CRYPTO_BYTES: usize = 32;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER76890S_AVX2_CRYPTO_SECRETKEYBYTES: usize = 2400;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER76890S_AVX2_CRYPTO_PUBLICKEYBYTES: usize = 1184;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER76890S_AVX2_CRYPTO_CIPHERTEXTBYTES: usize = 1088;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER76890S_AVX2_CRYPTO_BYTES: usize = 32;
 pub const PQCLEAN_KYBER102490S_CLEAN_CRYPTO_SECRETKEYBYTES: usize = 3168;
 pub const PQCLEAN_KYBER102490S_CLEAN_CRYPTO_PUBLICKEYBYTES: usize = 1568;
 pub const PQCLEAN_KYBER102490S_CLEAN_CRYPTO_CIPHERTEXTBYTES: usize = 1568;
 pub const PQCLEAN_KYBER102490S_CLEAN_CRYPTO_BYTES: usize = 32;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER102490S_AVX2_CRYPTO_SECRETKEYBYTES: usize = 3168;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER102490S_AVX2_CRYPTO_PUBLICKEYBYTES: usize = 1568;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER102490S_AVX2_CRYPTO_CIPHERTEXTBYTES: usize = 1568;
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 pub const PQCLEAN_KYBER102490S_AVX2_CRYPTO_BYTES: usize = 32;
 
 #[link(name = "kyber")]
@@ -96,13 +192,25 @@ extern "C" {
         sk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER512_AVX2_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER512_AVX2_crypto_kem_enc(ct: *mut u8, ss: *mut u8, pk: *const u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER512_AVX2_crypto_kem_dec(ss: *mut u8, ct: *const u8, sk: *const u8)
         -> c_int;
     pub fn PQCLEAN_KYBER768_CLEAN_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
@@ -113,13 +221,25 @@ extern "C" {
         sk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER768_AVX2_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER768_AVX2_crypto_kem_enc(ct: *mut u8, ss: *mut u8, pk: *const u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER768_AVX2_crypto_kem_dec(ss: *mut u8, ct: *const u8, sk: *const u8)
         -> c_int;
     pub fn PQCLEAN_KYBER1024_CLEAN_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
@@ -131,13 +251,25 @@ extern "C" {
         sk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER1024_AVX2_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER1024_AVX2_crypto_kem_enc(ct: *mut u8, ss: *mut u8, pk: *const u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER1024_AVX2_crypto_kem_dec(
         ss: *mut u8,
         ct: *const u8,
@@ -155,17 +287,29 @@ extern "C" {
         sk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER51290S_AVX2_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER51290S_AVX2_crypto_kem_enc(
         ct: *mut u8,
         ss: *mut u8,
         pk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER51290S_AVX2_crypto_kem_dec(
         ss: *mut u8,
         ct: *const u8,
@@ -183,17 +327,29 @@ extern "C" {
         sk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER76890S_AVX2_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER76890S_AVX2_crypto_kem_enc(
         ct: *mut u8,
         ss: *mut u8,
         pk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER76890S_AVX2_crypto_kem_dec(
         ss: *mut u8,
         ct: *const u8,
@@ -211,17 +367,29 @@ extern "C" {
         sk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER102490S_AVX2_crypto_kem_keypair(pk: *mut u8, sk: *mut u8) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER102490S_AVX2_crypto_kem_enc(
         ct: *mut u8,
         ss: *mut u8,
         pk: *const u8,
     ) -> c_int;
 
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     pub fn PQCLEAN_KYBER102490S_AVX2_crypto_kem_dec(
         ss: *mut u8,
         ct: *const u8,
@@ -263,7 +431,11 @@ mod test_kyber512_clean {
     }
 }
 #[cfg(test)]
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 mod test_kyber512_avx2 {
     use super::*;
 
@@ -334,7 +506,11 @@ mod test_kyber768_clean {
     }
 }
 #[cfg(test)]
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 mod test_kyber768_avx2 {
     use super::*;
 
@@ -405,7 +581,11 @@ mod test_kyber1024_clean {
     }
 }
 #[cfg(test)]
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 mod test_kyber1024_avx2 {
     use super::*;
 
@@ -480,7 +660,11 @@ mod test_kyber51290s_clean {
     }
 }
 #[cfg(test)]
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 mod test_kyber51290s_avx2 {
     use super::*;
 
@@ -555,7 +739,11 @@ mod test_kyber76890s_clean {
     }
 }
 #[cfg(test)]
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 mod test_kyber76890s_avx2 {
     use super::*;
 
@@ -630,7 +818,11 @@ mod test_kyber102490s_clean {
     }
 }
 #[cfg(test)]
-#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+#[cfg(all(
+    not(target_os = "windows"),
+    not(target_os = "macos"),
+    target_arch = "x86_64"
+))]
 mod test_kyber102490s_avx2 {
     use super::*;