+++ pqcrypto-kyber/src/kyber512.rs
@@ -101,7 +101,11 @@ pub const fn shared_secret_bytes() -> usize {
 
 /// Generate a kyber512 keypair
 pub fn keypair() -> (PublicKey, SecretKey) {
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     {
         if is_x86_feature_detected!("avx2") {
             return unsafe { keypair_avx2() };
@@ -136,7 +140,11 @@ unsafe fn keypair_avx2() -> (PublicKey, SecretKey) {
 
 /// Encapsulate to a kyber512 public key
 pub fn encapsulate(pk: &PublicKey) -> (SharedSecret, Ciphertext) {
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     {
         if is_x86_feature_detected!("avx2") {
             return unsafe { encapsulate_avx2(pk) };
@@ -185,7 +193,11 @@ unsafe fn encapsulate_avx2(pk: &PublicKey) -> (SharedSecret, Ciphertext) {
 
 /// Decapsulate the received kyber512 ciphertext
 pub fn decapsulate(ct: &Ciphertext, sk: &SecretKey) -> SharedSecret {
-    #[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
+    #[cfg(all(
+        not(target_os = "windows"),
+        not(target_os = "macos"),
+        target_arch = "x86_64"
+    ))]
     {
         if is_x86_feature_detected!("avx2") {
             return unsafe { decapsulate_avx2(ct, sk) };